requirements.txt
alembic==1.14.0
greenlet==3.1.1
Mako==1.3.8
MarkupSafe==3.0.2
Pygments==2.11.2
python-dotenv==1.0.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
ulid-py==1.1.0
